requirements.txt
pandas~=2.2.3
matplotlib~=3.9.2
Flask~=3.1.0
google-api-python-client~=2.155.0
numpy~=2.1.3
scikit-learn~=1.5.2
requests~=2.32.3
pillow~=11.0.0